Place tests alongside their code
