+++ lib/badge.spec.js
@@ -1,7 +1,7 @@
 const assert = require('assert');
 const isSvg = require('is-svg');
 
-const badge = require('../lib/badge');
+const badge = require('./badge');
 
 describe('The badge generator', function () {
   it('should produce SVG', function(done) {
