+++ lib/lru-cache.spec.js
@@ -1,6 +1,6 @@
 var assert = require('assert');
 
-var LRU = require('../lib/lru-cache.js');
+var LRU = require('./lru-cache');
 
 describe('The LRU cache', function () {
