+++ lib/svg-to-img.spec.js
@@ -2,8 +2,8 @@ const assert = require('assert');
 const sinon = require('sinon');
 const isPng = require('is-png');
 
-const badge = require('../lib/badge');
-const svg2img = require('../lib/svg-to-img.js');
+const badge = require('./badge');
+const svg2img = require('./svg-to-img');
 
 describe('The rasterizer', function () {
   let cacheGet;
